requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1
billiard==3.3.0.23
celery==3.1.26.post2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.7
django-celery==3.3.1
django-crontab==0.7.1
django-excel==0.0.10
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.1.0
importlib-metadata==3.4.0
kombu==3.0.37
lml==0.1.0
Markdown==3.3.3
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.2
prompt-toolkit==3.0.19
pyclean==2.0.0
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyexcel-xls==0.6.2
pyexcel-xlsx==0.6.0
PyMySQL==1.0.2
pytz==2021.1
PyYAML==5.4.1
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.1
sweetify==2.0.0
tablib==3.0.0
texttable==1.6.3
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
